+++ Source/ExtendedStorage/Patches/ITab_Storage_FillTab.cs
@@ -1,8 +1,12 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
 using System.Reflection;
 using System.Reflection.Emit;
 using Harmony;
 using RimWorld;
+using UnityEngine;
+using Verse;
 
 namespace ExtendedStorage
 {
@@ -10,6 +14,8 @@ namespace ExtendedStorage
     public class ITab_Storage_FillTab
     {
         private static readonly PropertyInfo piSelStoreSettingsParent;
+
+        public static bool showStoreSettings = false;
 
         static ITab_Storage_FillTab()
         {
@@ -41,7 +47,7 @@ namespace ExtendedStorage
         {
             ITab_Storage extended = tab as ITab_Storage;
 
-            if (extended == null)
+            if (extended == null || (DebugSettings.godMode && showStoreSettings))
             {
                 parent = GetSelStoreSettingsParent(tab);
                 return parent.GetStoreSettings();
@@ -50,6 +56,26 @@ namespace ExtendedStorage
 
             parent = building;
             return building?.userSettings;
+        }
+
+        /// <summary>
+        /// Add debug dropdown to toggle displayed settings to tab. Only visible in GodMode.
+        /// </summary>
+        public static void Postfix(RimWorld.ITab_Storage __instance) {
+            if (!DebugSettings.godMode || !(__instance is ITab_Storage))
+                return;
+
+            Rect rect = new Rect(160f+10f+5f, 10f, 100f, 29f);
+
+            Text.Font = GameFont.Tiny;
+            if (Widgets.ButtonText(rect, $"[Debug] {(showStoreSettings ? "Store" : "User")}", true, false, true)) {
+                List<FloatMenuOption> list = new List<FloatMenuOption>
+                                             {
+                                                 new FloatMenuOption("User", () => { showStoreSettings = false; }),
+                                                 new FloatMenuOption("Store", () => { showStoreSettings = true; })
+                                             };
+                Find.WindowStack.Add(new FloatMenu(list));
+            }
         }